requirements.txt
pandas>=1.5
openpyxl>=3.0
Pillow>=9.0
requests>=2.28
pytest>=7.0
